Use polymorphic nested inputs for copy specs
Previously we used the runtime TaskInputs API to register copy specs as inputs by listening for child specs being added. Once we wanted to know the actual files to copy, or the final state of inherited the properties, CopySpecResolvers were created to do that job. This code was hard to follow, somewhat redundant and inefficient in places.

This commit introduces the immutable ResolvedCopySpec and ResolvedCopySpecNode types that capture the resolved properties and sources used by both input snapshotting and the execution of the copy actions. The resolved root spec is exposed as a `@Nested` property on the AbstractCopyTask type.